Image converted to webp. 'Mixed Content' problem fixed. Old image deleting fixed.

+++ controllers/authControllers.js
@@ -11,6 +11,7 @@ import { unlink } from "fs/promises";
 import { nanoid } from "nanoid";
 import sendEmail from "../helpers/sendEmail.js";
 import ResetToken from "../models/ResetToken.js";
+import webp from 'webp-converter';
 
 const {
   JWT_SECRET,
@@ -177,46 +178,51 @@ const logout = async (req, res) => {
 };
 
 const updateUser = async (req, res) => {
-  const { _id: owner, password, avatarURL } = req.user;
-  const { new_password, password: old_password, email } = req.body;
-  const changedData = { ...req.body };
+  const {_id: owner, password, avatarURL} = req.user;
+  const {new_password, password: old_password, email} = req.body;
+  const changedData = {...req.body};
 
   if (old_password && new_password) {
     const user = await userServices.findUserById(owner);
     if (!user) {
-      throw HttpError(404, "User not found!");
+      throw HttpError(404, 'User not found!');
     }
 
     const passwordCompare = await bcrypt.compare(old_password, password);
     if (!passwordCompare) {
-      throw HttpError(401, "Incorrect password!");
+      throw HttpError(401, 'Incorrect password!');
     }
 
     changedData.password = await bcrypt.hash(new_password, 8);
   }
 
-  if (req.file) {
-    const { path: filePath } = req.file;
-    Jimp.read(filePath)
-      .then((image) => {
-        return image.cover(250, 250).quality(60).write(filePath);
-      })
-      .catch((err) => {
-        console.error(err);
-      });
+  const user = await userServices.findUser({email});
+  if (user) {
+    throw HttpError(409, 'Email is already used');
+  }
 
-    const { url: photo } = await cloudinary.uploader.upload(filePath, {
-      folder: "avatars",
+  if (req.file) {
+    const {path: filePath} = req.file;
+    const webpPath = `${filePath.split('.')[0]}.webp`;
+    const image = await Jimp.read(filePath);
+    await image.cover(250, 250).write(filePath);
+    await webp.cwebp(filePath, webpPath, '-q 100');
+    let {url: photo} = await cloudinary.uploader.upload(webpPath, {
+      folder: 'avatars'
     });
+
+    photo = photo.replace('http', 'https')
 
     changedData.avatarURL = photo;
     await unlink(filePath);
+    await unlink(webpPath);
   }
 
   const newUser = await userServices.updateUser(owner, changedData);
 
   if (req.file && newUser) {
-    await cloudinary.uploader.destroy(avatarURL);
+    const avatar_id = avatarURL.split('/').pop().split('.')[0]
+    await cloudinary.uploader.destroy(`avatars/${avatar_id}`);
   }
 
   res.status(201).json({
@@ -224,8 +230,8 @@ const updateUser = async (req, res) => {
       email: newUser.email,
       username: newUser.username,
       gender: newUser.gender,
-      avatarURL: newUser.avatarURL,
-    },
+      avatarURL: newUser.avatarURL
+    }
   });
 };
